#7 Set comment character based on users config
For commit message hooks this will set the comment character based on
what the user has had set in their git config settings. This means that
if a user wants to be able, for example put a pull request ID (like in
this commit messages subject) at the start of their commit message,
they can.