Fix prepare-commit-msg hook
To make sure that binary (cli) and php hooks work together the
repository and the application have to be in sync.

The user should not have to worry about this, so every php
prepare-msg-hook action has to write the changed commit message to
disk after every action.

To archive this the pre and post actions where moved from the command
to the actual hook runner. The runner is now able to execute tasks
between each action as well as once before all actions and once after
all actions are executed.

Have a look at:
Hook::beforeHook
Hook::beforeAction
Hook::afterAction
Hook::afterHook

To handle this differently for each hook the hook runner
became abstract and is 'replaced' with specific runner implementations
for each hook.