Grant access to original hook arguments
In order to implement useful cli hooks those hooks must have access to
the original hook arguments provided by git.

In order to do this you now have the possibility to add placeholder
to your cli commands for example to the commit-msg hook like this.

  /my/command --commit-msg-file={FILE}

The {FILE} part will be replaced by captain hook with the original
argument provided by git.

For a list of available replacements have a look at

  \CaptainHook\App\Runner\Action\Cli::formatCommand