Add possibility to use ActionFactory
This adds the possibiliity to use an ActionFactory in the configuration
instead of an Action directly.

The Options from the configuration are passed to the
ActionFactory::getAction-Method to enable creation of customized
classes. these classes need to implement the Action-Interface.

That way the creation of the custom-class is separated from the actual
execution of the Action.