Add 'Debug' action
This action can be used to output certain hook information. If you use
this CaptainHook will always fail because by default it throws an
exception. It can be used to test message validation or other
pre-commit stuff without executing the commit. This is helpful if
you develop your own Action or want to verify your configuration
is working.