Add 'conditions' feature
Now you can add 'conditions' to action configurations. Conditions
can be PHP classes implementing the Condition interface or CLI scripts
that have an exit code of 0 if the condition applies and an exit code
other then 0 when die condition does not apply.

Multiple Conditions for one action are combined via 'and' so all
Conditions for an Action must apply to execute the Action.

Issue #27