Add docker run-mode
Now there are different templates that generate different hook
sourcecode for different use cases. There's a template to create
the standard php hooks that work if you have installed php locally.
And there is a template to create bash hooks that startup a docker
container to execute CaptainHook inside a container.

If you want to use the docker variant you have to install your hooks
like this

$ captainhook install --run-mode=docker --container=captainhook-php7.3