Change Docker --container to --command
As discussed in issue #48 the --container option was replaced with the
--command option. This breaks the current docker implementation.
Since the docker feature is not documented and still considered
alpha this is not considered a BC break.

This change allows for more flexibility on spinning up docker
containers. You are not bound to 'docker exec' anymore, you can no use
'docker run' or 'docker-compose run' as well.