Fix docker-setup with configuration file
This commit fixes that the setup reads the docker-setup configuration
from the captainhook.json config file.

Currently the docker config can only be setup using the commandline
arguments on the `captainhook install` command. But for existing or new
setups where the configuration file is distributed that is not the
expected behaviour. The problem was that the Builder currently only
checks for the command-line options but ignores the configuration
option. The new behaviour also checks the config options and uses them
when no commandline arguments are passed. Therefore installations using
the captainhook composer plugin now will be able to install the docker
based setup from the commited config-file