Add special cli command placeholder
As requested a while ago in issue #52 it would be nice to have
placeholders for commands that get replaced before executing the
command.

The original hook arguments have been available as placeholders for
quite some time. This commit adds the {$STAGES_FILES} placeholder.

You can use use it to pass all new or changed files within the
current commit to a command like phpcs.