Add move existing hooks feature
You can now move existing hooks to a specified directory during
hook installation. This way you not just blindly overwrite existing
hooks.

For example
  captainhook install --move-existing-to=hooks-old

This can be useful if you are using git-lfs to move the git-lfs
hooks to another location to later include them in the
captainhook.json and let the Captain execute them.