Add configurable Or and And conditions
This allows one to combine conditions for when to run git hooks using
boolean AND as well as boolean OR via the configuration.

The current conditional configuration is extended to use "and" or "or"
as value for the "exec" key and then an array of
condition-configurations as value for the "args" key.
So an example configuration would look like this:

```
"conditions": [{
    "exec": "and",
    "args": [{
        "exec":
"\\CaptainHook\\App\\Hook\\Condition\\AnyFileChanged",
        "args": [
            ["foo.php", "bar.php"]
        ]
    },{
        "exec":
"\\CaptainHook\\App\\Hook\\Condition\\AnyFileChanged",
        "args": [
            ["foo.php", "bar.php"]
        ]
    }]
}
```