Add IntegrateBeforePush action
The action monitors a specific remote branch and forces you to
integrate changes from that monitored branch into our local branch
if a configured string is appearing in a commit message in the
monitored branch.

For example if you made changes to the build system and in order to
make it work every feature branch has to get the latest settings to
not uselessly fail builds.