Catch condition exceptions
In order to properly output any errors occurring during condition
execution the skip handling was moved into the try catch.

This leads to Plugin::afterAction being called even if the condition
fails, that was not happening before and may cause some plugins to
behave differently in case of faulty conditions.