Prevent PHAR from crashing unnecessarily
If no autoloader is set the PHAR file does not need a bootstrap file.
As long as you are only using Cap'n functionality you are good to go.
This fix makes sure the Cap'n only returns an error if a bootstrap
file is not available that is set explicitly.

Fix issue #248